+++ controllers/resolver_rules.go
@@ -119,7 +119,7 @@ func (r *ResolverRulesReconciler) reconcileNormal(ctx context.Context, awsCluste
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	cluster := buildClusterFromAWSCluster(awsCluster, identity)
+	cluster := buildClusterFromAWSCluster(awsCluster, identity, nil)
 
 	awsAccountOwnerOfRulesToAssociate, ok := awsCluster.Annotations[gsannotations.ResolverRulesOwnerAWSAccountId]
 	if !ok {
@@ -150,7 +150,7 @@ func (r *ResolverRulesReconciler) reconcileNormal(ctx context.Context, awsCluste
 func (r *ResolverRulesReconciler) reconcileDelete(ctx context.Context, awsCluster *capa.AWSCluster, identity *capa.AWSClusterRoleIdentity) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 
-	cluster := buildClusterFromAWSCluster(awsCluster, identity)
+	cluster := buildClusterFromAWSCluster(awsCluster, identity, nil)
 
 	err := r.resolver.DeleteRule(ctx, logger, cluster)
 	if err != nil {